Get rid of contacts directory and move vehicles to lib
