+++ rmndin/rmndin/lib/vehicles.py
