+++ rmndin/rmndin/reminders/models.py
@@ -2,8 +2,6 @@ from rmndin import db
 from rmndin.lib.db.mixins import BaseMixin
 from rmndin.lib.db.enums import DeliveryMethodEnum
 
-from rmndin.contacts.vehicles import RedditContactVehicle
-
 
 class Reminder(BaseMixin, db.Model):
     id = db.Column(db.Integer, primary_key=True)
@@ -15,15 +13,3 @@ class Reminder(BaseMixin, db.Model):
     delivery_method = db.Column(db.Enum(DeliveryMethodEnum), nullable=False)
 
     fulfilled = db.Column(db.String, nullable=False, default=False)
-
-    def get_vehicle(self):
-        if self.method == DeliveryMethodEnum.reddit:
-            veh = RedditContactVehicle(user_contact=self,
-                                       verified=self.verified)
-        # elif self.method == DeliveryMethodEnum.email:
-        #     veh = EmailVehicle(email=self.identifier,
-        #                        verified=self.verified)
-        else:
-            veh = None
-
-        return veh
